Fixed package not working on non-English systems
Because Tidy only offers error output as lines of human-readable strings, this package uses a regular expression to parse each line. Unfortunately, this means that if the locale changes to use a non-English language, the regular expression breaks.

This commit fixes the issue in the short term by forcing Tidy to use English via an executable argument.